fixes for user

+++ projects/views.py
@@ -50,6 +50,7 @@ class ProjectDetail(APIView):
     
     def put(self, request, pk):
         project = self.get_object(pk)
+        self.check_object_permissions(request, project)
         data = request.data
         serializer = ProjectDetailSerializer(
             instance=project,
@@ -61,6 +62,7 @@ class ProjectDetail(APIView):
 
     def delete(self, request,pk):
         project = self.get_object(pk)
+        self.check_object_permissions(request, project)
         project.delete()
         return Response(status.HTTP_204_NO_CONTENT)
